+++ src/iotvm-eventengine/src/main/java/org/softwareforce/iotvm/eventengine/cep/fabrication/ExponentialSmoothingWithLinearTrendOptimization.java
@@ -1,6 +1,9 @@
 package org.softwareforce.iotvm.eventengine.cep.fabrication;
 
+import com.google.common.base.Preconditions;
 import java.util.List;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.softwareforce.iotvm.eventengine.cep.CalculationUtils;
 
 /**
@@ -12,67 +15,132 @@ import org.softwareforce.iotvm.eventengine.cep.CalculationUtils;
 public class ExponentialSmoothingWithLinearTrendOptimization
     extends ExponentialSmoothingWithLinearTrend {
 
-  private double bestAlpha;
-  private double bestBeta;
-  private double bestMSE;
+  private static final Logger LOGGER =
+      LoggerFactory.getLogger(ExponentialSmoothingWithLinearTrendOptimization.class);
+
+  private double selectedAlpha;
+  private double selectedBeta;
+  private double MSE;
+  private double MAE;
   private List<Double> bestForecasts;
+  private boolean debug;
 
   public ExponentialSmoothingWithLinearTrendOptimization(double alpha, double beta, int horizon) {
     super(alpha, beta, horizon);
     this.reset();
+    this.debug = false;
   }
 
   private void reset() {
-    this.bestAlpha = 0;
-    this.bestBeta = 0;
-    this.bestMSE = Double.MAX_VALUE;
+    this.selectedAlpha = 0;
+    this.selectedBeta = 0;
+    this.MSE = Double.MAX_VALUE;
+    this.MAE = Double.MAX_VALUE;
     this.bestForecasts = null;
+    // TODO Idea: Do not trust forecast if MAE is not in a specific range.
   }
 
-  public double getBestAlpha() {
-    return this.bestAlpha;
+  public double getSelectedAlpha() {
+    return this.selectedAlpha;
   }
 
-  public double getBestBeta() {
-    return this.bestBeta;
+  public double getSelectedBeta() {
+    return this.selectedBeta;
   }
 
-  public double getBestMSE() {
-    return this.bestMSE;
+  public double getMSE() {
+    return this.MSE;
+  }
+
+  public double getMAE() {
+    return this.MAE;
   }
 
   public List<Double> getBestForecasts() {
     return this.bestForecasts;
   }
 
+  public boolean getDebug() {
+    return this.debug;
+  }
+
+  public void setDebug(boolean debug) {
+    this.debug = debug;
+  }
+
   public void optimizeParameters(List<Double> data, double[] alphaRange, double[] betaRange) {
     this.reset();
 
-    for (double alpha : alphaRange) {
-      for (double beta : betaRange) {
+    final long startNs = System.nanoTime();
+
+    int iterationCount = 0;
+    for (final double alpha : alphaRange) {
+      for (final double beta : betaRange) {
         this.setAlpha(alpha);
         this.setBeta(beta);
 
         final List<Double> forecast = this.forecast(data);
-        if (data.size() != forecast.size() - this.getHorizon()) {
-          throw new IllegalStateException();
-        }
+        Preconditions.checkState(data.size() == (forecast.size() - this.getHorizon()));
 
         final List<Double> yTrue = data.subList(0, data.size());
         // yPred (in sample predictions)
         final List<Double> yPred = forecast.subList(0, data.size());
-        if (yTrue.size() != yPred.size()) {
-          throw new IllegalStateException();
+        Preconditions.checkState(yTrue.size() == yPred.size());
+
+        double MSE = CalculationUtils.calculateMSE(yTrue, yPred);
+        double MAE = CalculationUtils.calculateMAE(yTrue, yPred);
+
+        boolean isBest = false;
+        if (MAE < this.MAE) {
+          isBest = true;
+        } else if (MAE == this.MAE) {
+          if (MSE < this.MSE) {
+            isBest = true;
+          }
         }
 
-        double mse = CalculationUtils.calculateMSE(yTrue, yPred);
-        if (mse < this.bestMSE) {
-          this.bestMSE = mse;
-          this.bestAlpha = alpha;
-          this.bestBeta = beta;
+        if (isBest) {
+          this.selectedAlpha = alpha;
+          this.selectedBeta = beta;
+          this.MSE = MSE;
+          this.MAE = MAE;
           this.bestForecasts = forecast;
         }
+
+        iterationCount++;
       }
     }
+
+    if (!this.debug) {
+      return;
+    }
+
+    final long endNs = System.nanoTime();
+    final long diffNs = endNs - startNs;
+    final long diffMs = diffNs / 1000000;
+
+    final List<Double> bestForecastsInHorizon =
+        this.bestForecasts.subList(data.size(), this.bestForecasts.size());
+    Preconditions.checkState(bestForecastsInHorizon.size() == this.getHorizon());
+    LOGGER.info(
+        "Completed parameter optimization ({} iterations, {} ms) with MSE {}, MAE {}, alpha {},"
+            + " beta {}, last real {}, last forecast {}",
+        iterationCount,
+        diffMs,
+        this.MSE,
+        this.MAE,
+        this.selectedAlpha,
+        this.selectedBeta,
+        data.getLast(),
+        this.bestForecasts.getLast());
+  }
+
+  public static double[] generateRange(double start, double end, double step) {
+    final int size = (int) ((end - start) / step) + 1;
+    final double[] array = new double[size];
+    for (int i = 0; i < size; i++) {
+      array[i] = start + (i * step);
+    }
+    return array;
   }
 }